+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+ * added driver version to APIException
+
 ## 1.2.3
 
  * replaced deprecated crosswalk query with regular table read